LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/getclientrects-007-expected.txt


FAIL #col1 assert_equals: expected 3 but got 1
FAIL #col2 assert_equals: expected 3 but got 1
FAIL #col3 assert_equals: expected 3 but got 1
FAIL #col4 assert_equals: expected 3 but got 1
FAIL #group assert_equals: expected 3 but got 1
